pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kube-netpol"
version = "1.0.0"
description = "Kubernetes NetworkPolicy generator, validator & visualizer"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Sai Sandeep", email = "ssan@example.com"}
]
keywords = ["kubernetes", "networkpolicy", "network", "security", "firewall", "visualizer", "generator"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: System :: Networking",
    "Topic :: System :: Systems Administration",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
]
dependencies = [
    "click>=8.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/SanjaySundarMurthy/kube-netpol"
Repository = "https://github.com/SanjaySundarMurthy/kube-netpol"
Issues = "https://github.com/SanjaySundarMurthy/kube-netpol/issues"

[project.scripts]
kube-netpol = "kube_netpol.cli:main"

[tool.setuptools.packages.find]
include = ["kube_netpol*"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py39"
line-length = 200

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
